BasketOrderManager is Added

+++ sumplierapp/sumplierapp/Model/CustomerProduct.cs
@@ -14,12 +14,12 @@ namespace sumplierapp.Model
         public long ProductCode { get; set; }
         public string Barcode { get; set; }
         public string ProductName { get; set; }
-        public double Price { get; set; }
-        public double Price2 { get; set; }
-        public double Price3 { get; set; }
-        public double TaxRate { get; set; }
-        public double TaxRate2 { get; set; }
-        public double TaxRate3 { get; set; }
+        public decimal Price { get; set; }
+        public decimal Price2 { get; set; }
+        public decimal Price3 { get; set; }
+        public decimal TaxRate { get; set; }
+        public decimal TaxRate2 { get; set; }
+        public decimal TaxRate3 { get; set; }
         public string Image { get; set; }
         public string Description { get; set; }
         public bool IsActive { get; set; }
